Changes to make it all work
n the package there was some "crap" that caused errors (<<<<<<< HEAD and
stuff like that). I just removed them. In tweety.js the references to
convertor.js and tweetcipher.js were missing the /lib folder (and at
least in Windows it caused trouble (the add on would start, but it
wouldn't display the button.